refactor(item): extract CLI business logic into services/items
Move query-building, validation, and column-value rendering out of
src/mondo/cli/item.py into a new src/mondo/services/items.py. The Typer
callbacks keep argument parsing, emission, polling, and exit-code mapping;
the service holds pure helpers (split_filter_expr, build_filter_rule,
parse_columns_csv, build_query_params, parse_column_mapping,
read_batch_input, can_slim_column_values, PositionRelative) and
client-backed helpers (fetch_column_defs, build_column_values,
build_create_variables_for_row, resolve_item_target).

Behavior-neutral: function bodies moved verbatim (renamed _private ->
public), the CLI keeps its try/except -> exit-code mapping, and
GlobalOpts/MondayClient are TYPE_CHECKING-only imports to avoid a cycle.
Drop imports orphaned by the move and fix two doc-comment cross-references
(output/fields.py, subitem.py) that pointed at the relocated symbols.

item.py: 1320 -> 1027 lines. Full unit suite (1639) green; ruff/mypy clean.
